+++ xmake/core/sandbox/modules/os.lua
@@ -58,6 +58,7 @@ sandbox_os.addenv       = os.addenv
 sandbox_os.setenvp      = os.setenvp
 sandbox_os.addenvp      = os.addenvp
 sandbox_os.getenvs      = os.getenvs
+sandbox_os.setenvs      = os.setenvs
 sandbox_os.pbpaste      = os.pbpaste
 sandbox_os.pbcopy       = os.pbcopy
 sandbox_os.cpuinfo      = os.cpuinfo